Spotify Streamer: Stage 2. All required plus all optional components implemented. Tested on phone and tables and works correctly, if you have any problem please contact me, I'm not used to Git and I may forgot to upload something.
* Now playing feature is extended so it can also resume the playback queue from the last session (last execution of the app)

* Share option feature is extended so it can also share the artist (with spotify's artist url)

* I did not mention this on Stage 1, but I've also implemented an artist image list / track image list view to see results as swipeable and zoomeable images.
It shows up when any artist image or track image is clicked on the results lists.
The view also has the ability to share the current artist or song, and if the image is clicked, to start the top tracks search (on artist mode) or the track player (on track mode).

* Player has a "Continuous play" button that when enabled plays all tracks continuously. This option button is stored in preferences.
Current song image can also be zoomed, and if swiped (left or right), the next or previous song is played.